Split "user" module
