+++ src/store/mutation-types.js
@@ -45,3 +45,4 @@ export const DESTROY_GAME = 'DESTROY_GAME'
 export const SHOW_ERROR = 'SHOW_ERROR'
 export const HIDE_ERROR = 'HIDE_ERROR'
 export const HIDE_ALL_ERRORS = 'HIDE_ALL_ERRORS'
+export const HIDE_ERRORS_OF_TYPE = 'HIDE_ERRORS_OF_TYPE'
